third_party/WebKit/LayoutTests/http/tests/security/xssAuditor/script-tag-with-source-same-host-with-query-expected.txt

CONSOLE ERROR: line 4: The XSS Auditor refused to execute a script in 'http://localhost:8000/security/xssAuditor/resources/echo-intertag.pl?q=%3Cscript%20src=%27xss.js?maybe+dangerous+query+string%27%3E%3C/script%3E' because its source code was found within the request. The server sent an 'X-XSS-Protection' header requesting this behavior.
